+++ tests/Traits/Helpers/NestedArrayAccessorTest.php
@@ -1,12 +1,15 @@
 <?php
 
-namespace Helpers;
+namespace Traits\Helpers;
 
 use PHPUnit\Framework\TestCase;
 use UQL\Helpers\ArrayHelper;
+use UQL\Traits\Helpers\NestedArrayAccessor;
 
-class ArrayHelperTest extends TestCase
+class NestedArrayAccessorTest extends TestCase
 {
+    use NestedArrayAccessor;
+
     private array $testArray = [
         'a' => [
             'b' => [
@@ -23,26 +26,26 @@ class ArrayHelperTest extends TestCase
 
     public function testGetNestedValue(): void
     {
-        $this->assertEquals('value', ArrayHelper::getNestedValue($this->testArray, 'a.b.c'));
-        $this->assertEquals('value2', ArrayHelper::getNestedValue($this->testArray, 'a.d'));
+        $this->assertEquals('value', $this->accessNestedValue($this->testArray, 'a.b.c'));
+        $this->assertEquals('value2', $this->accessNestedValue($this->testArray, 'a.d'));
 
-        $result = ArrayHelper::getNestedValue($this->testArray, 'a.e[]->x');
+        $result = $this->accessNestedValue($this->testArray, 'a.e[]->x');
         $this->assertEquals(1, $result[0]);
         $this->assertEquals(2, $result[1]);
         $this->assertEquals(3, $result[2]);
 
-        $result = ArrayHelper::getNestedValue($this->testArray, 'a.e[]->y');
+        $result = $this->accessNestedValue($this->testArray, 'a.e[]->y');
         $this->assertEquals(2, $result[0]);
         $this->assertEquals(3, $result[1]);
         $this->assertEquals(4, $result[2]);
 
-        $result = ArrayHelper::getNestedValue($this->testArray, 'a.e[]->z');
+        $result = $this->accessNestedValue($this->testArray, 'a.e[]->z');
         $this->assertEquals(3, $result[0]);
         $this->assertEquals(4, $result[1]);
         $this->assertEquals(5, $result[2]);
 
-        $this->assertEquals(3, ArrayHelper::getNestedValue($this->testArray, 'a.e.0->z'));
-        $this->assertEquals(4, ArrayHelper::getNestedValue($this->testArray, 'a.e.1->z'));
-        $this->assertEquals(5, ArrayHelper::getNestedValue($this->testArray, 'a.e.2->z'));
+        $this->assertEquals(3, $this->accessNestedValue($this->testArray, 'a.e.0->z'));
+        $this->assertEquals(4, $this->accessNestedValue($this->testArray, 'a.e.1->z'));
+        $this->assertEquals(5, $this->accessNestedValue($this->testArray, 'a.e.2->z'));
     }
 }